Add service to extract infos from Jaeger
We want to have the personal information, which is gathered be Jaeger,
extracted from Jaeger. Therefore I add a simple json server, which
currently extracts the purposes for all services. It also has a simple
index page desribing itself.

This service relies on the jaeger query service and will report an error
in json, if that service is not reachable.
The host for jaeger-query can be set via the environment variable
'JAEGER_QUERY_HOST'.

Close #34