Cleanup: Drop no longer used function
We thought it might be a good idea to have a function, which can set all
the information in 1 place. We think that's actually not really going to
be used and it will also look probably not so nice with the upcoming
generics. Therefore I remove it now.

I will mention this commit in the documentation to note this
alternative function.